Fix androidTests not passing due to previous refactoring
