+++ app/src/main/java/ch/epfl/favo/view/tabs/Tab2.java
@@ -52,6 +52,7 @@ public class Tab2 extends Fragment implements View.OnClickListener {
     // Replace the current fragment with the new fragment.
     // Todo: Seems useful. Try to put this method in a util package and import it here.
     private void replaceFragment(Fragment newFragment) {
+        assert getFragmentManager() != null;
         FragmentTransaction transaction = getFragmentManager().beginTransaction();
         transaction.replace(R.id.fragment_tab2, newFragment);
         transaction.addToBackStack(null);
